Move new consensus model to temporary package, see #115
